Moved trimTrailingValues to Utils
