+++ apps/balancer-tools/src/app/poolsimulator/(workers)/impact-curve-calculation.ts
@@ -2,12 +2,12 @@ import { AMM } from "@bleu-balancer-tools/math-poolsimulator/src";
 import { PoolPairData } from "@bleu-balancer-tools/math-poolsimulator/src/types";
 
 import { AnalysisData } from "#/contexts/PoolSimulatorContext";
+import { trimTrailingValues } from "#/lib/utils";
 
 import { PoolTypeEnum, TokensData } from "../(types)";
 import {
   calculateCurvePoints,
   convertAnalysisDataToAMM,
-  trimTrailingValues,
 } from "../(utils)";
 
 export interface ImpactWorkerInputData {
